Move tags into a dropdown
